Refactor one-off components into PaginationToolbar
It’s the only place these are used, so this removes the abstraction.

- Refactors `documents/Paginator` into `PaginationToolbar`
- Refactors `documents/Sections` into `PaginationToolbar`
